+++ sources/tools/SiliconStudio.Paradox.Importer.FBX/MeshConverter.cpp
@@ -613,7 +613,7 @@ public:
 	}
 
 	// return a boolean indicating whether the built material is transparent or not
-	MaterialAsset^ ProcessMeshMaterialAsset(FbxSurfaceMaterial* lMaterial, std::map<std::string, int>& uvElementMapping)
+	MaterialAsset^ ProcessMeshMaterialAsset(FbxSurfaceMaterial* lMaterial, std::map<std::string, size_t>& uvElementMapping)
 	{
 		auto uvEltMappingOverride = uvElementMapping;
 		auto textureMap = gcnew Dictionary<IntPtr, ComputeTextureColor^>();
@@ -936,7 +936,7 @@ public:
 		return false;
 	}
 
-	IComputeNode^ GenerateSurfaceTextureTree(FbxSurfaceMaterial* lMaterial, std::map<std::string, int>& uvElementMapping, Dictionary<IntPtr, ComputeTextureColor^>^ textureMap,
+	IComputeNode^ GenerateSurfaceTextureTree(FbxSurfaceMaterial* lMaterial, std::map<std::string, size_t>& uvElementMapping, Dictionary<IntPtr, ComputeTextureColor^>^ textureMap,
 												std::map<std::string, int>& textureNameCount, char const* surfaceMaterial, char const* surfaceMaterialFactor,
 												SiliconStudio::Paradox::Assets::Materials::MaterialAsset^ finalMaterial)
 	{
@@ -1137,7 +1137,7 @@ public:
 		return fileNameToUse;
 	}
 
-	ComputeTextureColor^ GenerateMaterialTextureNodeFBX(FbxFileTexture* lFileTexture, std::map<std::string, int>& uvElementMapping, Dictionary<IntPtr, ComputeTextureColor^>^ textureMap, std::map<std::string, int>& textureNameCount, SiliconStudio::Paradox::Assets::Materials::MaterialAsset^ finalMaterial)
+	ComputeTextureColor^ GenerateMaterialTextureNodeFBX(FbxFileTexture* lFileTexture, std::map<std::string, size_t>& uvElementMapping, Dictionary<IntPtr, ComputeTextureColor^>^ textureMap, std::map<std::string, int>& textureNameCount, SiliconStudio::Paradox::Assets::Materials::MaterialAsset^ finalMaterial)
 	{
 		auto texScale = lFileTexture->GetUVScaling();		
 		auto texturePath = FindFilePath(lFileTexture);
@@ -1590,7 +1590,7 @@ private:
 		}
 	}
 
-	MaterialInstantiation^ GetOrCreateMaterial(FbxSurfaceMaterial* lMaterial, List<String^>^ uvNames, List<MaterialInstantiation^>^ instances, std::map<std::string, int>& uvElements, std::map<FbxSurfaceMaterial*, std::string>& materialNames)
+	MaterialInstantiation^ GetOrCreateMaterial(FbxSurfaceMaterial* lMaterial, List<String^>^ uvNames, List<MaterialInstantiation^>^ instances, std::map<std::string, size_t>& uvElements, std::map<FbxSurfaceMaterial*, std::string>& materialNames)
 	{
 		for (int i = 0; i < instances->Count; ++i)
 		{
@@ -1665,7 +1665,7 @@ private:
 				if (lMaterialElement != NULL)
 				{
 					FbxSurfaceMaterial* lMaterial = pNode->GetMaterial(i);
-					std::map<std::string, int> uvElements;
+					std::map<std::string, size_t> uvElements;
 					auto uvNames = gcnew List<String^>();
 					for (int j = 0; j < pMesh->GetElementUVCount(); ++j)
 					{
@@ -1707,7 +1707,7 @@ private:
 		auto materials = gcnew Dictionary<String^, MaterialAsset^>();
 		for (int i = 0;  i < scene->GetMaterialCount(); i++)
 		{
-			std::map<std::string, int> dict;
+			std::map<std::string, size_t> dict;
 			auto lMaterial = scene->GetMaterial(i);
 			auto materialName = materialNames[lMaterial];
 			materials->Add(gcnew String(materialName.c_str()), ProcessMeshMaterialAsset(lMaterial, dict));
